[AIMRefund] Pass credit card expiration date only when an expiration was set
Authorize.Net does not require the expiration date when performing refunds. When no expiration is passed in and a refund request is made, the response comes back with an error saying the card has expired. 

This check makes it so the expiration date is only sent when a valid expiry date is passed in to the request parameters. 

> At least the last four digits of the credit card number (x_card_num) used for the
original, successfully settled transaction are submitted. An expiration date is not
required.